refactor(css): hide internal preprocessor types and expose types used for options
