+++ packages/vite/package.json
@@ -40,6 +40,7 @@
     "./types/*": {
       "types": "./types/*"
     },
+    "./types/internal/*": null,
     "./package.json": "./package.json"
   },
   "typesVersions": {
