+++ packages/vite/types/internal/cssPreprocessorOptions.d.ts
